+++ Semesterprojekt/src/worldofzuul/entities/Entity.java
@@ -7,10 +7,10 @@ import java.awt.Graphics;
  * @author Group 7
  */
 public abstract class Entity {
-    protected float x, y;
+    protected int x, y;
     protected int width, height;
     
-    public Entity(float x, float y, int width, int height) {
+    public Entity(int x, int y, int width, int height) {
         this.x = x;
         this.y = y;
         this.width = width;
@@ -23,19 +23,19 @@ public abstract class Entity {
     
     // GETTERS & SETTERS
     
-    public float getX() {
+    public int getX() {
         return x;
     }
 
-    public void setX(float x) {
+    public void setX(int x) {
         this.x = x;
     }
 
-    public float getY() {
+    public int getY() {
         return y;
     }
 
-    public void setY(float y) {
+    public void setY(int y) {
         this.y = y;
     }
